Ensures the Threads Are Running
This removes the need to restart the resource. By placing the threads into a function and calling the function at resource start and at the time the player loads into the server. The threads are guaranteed to be running.

The server counts resource start when the server starts the resource not when the client loads the resource.

That's why you place it into on playerloaded and into onresource start so when the player is in the game the function fires and when the server reloads the function fires again while the player is in the game.

Cheeers!